refactor: reduce duplication — share server/src values via common/ (#826)
jscpd flagged 4 duplicate-code alerts; a scan at min-tokens 30 showed the real
pattern behind them: values and wire types hand-mirrored between server/ and
src/, even though common/ is included by BOTH tsconfigs and already carries
dirChrome / themeColors / modelPresets.

The mirroring was self-perpetuating: firebase.ts and firebaseConfig.ts each
carried a comment claiming "the server tsconfig does not cross into src/, so
MulmoTerminal mirrors small shared values rather than importing a shared
module" — which has not been true since common/ existed.

Moved to common/ (each was a verbatim copy, now one definition):
- firebaseConfig, Shortcut/ShortcutKind/sameShortcut, GitStatus,
  LaunchProviderOption/LaunchOptions, GhItemBase/PrItem/RepoPrs/IssueItem/RepoIssues
- EMPTY_DIR_CHROME, so both DirConfig empties spread one base

The 4 alerts:
- TerminalCell rendered the expand/close pair twice; both now use the existing
  CellChromeButtons. Its `.stop` was redundant — shouldZoomOnHeaderClick already
  ignores clicks inside a button — and "close" stays an emit, so TerminalCell
  keeps its own confirming handler.
- AppRouteDeps / HookDeps restated the same 7 callbacks: now one
  SessionActivityDeps both extend.
- TEXT_EXTS / IN_APP_EXTENSIONS share 45 of their entries but answer different
  questions, so only the intersection moved to SOURCE_CODE_EXTENSIONS; each
  side keeps its own extras. Both resulting sets are byte-identical to before.
- SessionTabBar / Sidebar: folded isUnread into useSessionFilter. What remains
  is defineProps/defineEmits boilerplate Vue requires per component.

Behaviour-preserving throughout. test/common/sourceExtensions.spec.ts pins the
deliberate asymmetries (.md routes to the rendered viewer, .txt opens in Files,
dotfiles are server-only) so the next reader can't "fix" them into symmetry;
verified failing against three mutations of the split.
