+++ src/config/firebase.ts
@@ -6,7 +6,7 @@
 import { initializeApp } from "firebase/app";
 import { getAuth } from "firebase/auth";
 
-import { firebaseConfig } from "./firebaseConfig";
+import { firebaseConfig } from "../../common/firebaseConfig";
 
 export const firebaseApp = initializeApp(firebaseConfig);
 export const auth = getAuth(firebaseApp);
